Conn.handler: Refactor code to handle nonblocking TCP connect
On PAL/Linux the pal_connect() is now actually returning the
PAL_ERR_SOCKET_IN_PROGRES which we need to handle. Refactor
code a bit to handle that.

Note: the code is now tested on Linux only and this may fail on
mbed-os as long as IOTPAL-228 -issue is open.