connectionhandler: silence the compiler warnings from trace outputs
GCC complained rightfully for trace prints where a palStatus_t or
uint32_t were printed with "%d". Change PAL value prints to be
casted to int and use PRIu32 for the uint32_t.

While here, add the ESocketStateConnectBeingCalled to switch
case in dns_handler() to silence a warning about it.